requirements.txt
python-dotenv==0.21.1
transitions==0.9.0
vk-api==11.9.9
